brain-tumor-detection/backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
torch>=2.2.0
torchvision>=0.15.2
pillow==10.0.0
numpy>=1.26.0
